requirements.txt
astroid==2.1.0
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
Click==7.0
colorama==0.4.1
dj-database-url==0.5.0
Django==2.2
django-filter==2.1.0
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
elasticsearch==6.3.1
Flask==1.0.3
gunicorn==19.8.1
idna==2.6
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.3.1
mailjet-rest==1.3.0
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.2
PyJWT==1.7.1
pylint==2.2.2
python-dotenv==0.10.2
pytz==2018.4
requests==2.18.4
six==1.12.0
sqlparse==0.3.0
typed-ast==1.1.1
urllib3==1.22
Werkzeug==0.15.4
whitenoise==4.1.2
wrapt==1.10.11
